:truck: Move files to match standard neovim plugin architecture.
